src/main/java/by/ereut/balanceservice/client/ClientController.java
package by.ereut.balanceservice.client;

import lombok.extern.slf4j.Slf4j;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/clients")
@Slf4j
public class ClientController {

    private final ClientService service;

    public ClientController(ClientService service) {
        this.service = service;
    }

    @GetMapping(produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
    public List<Client> getAll(@RequestHeader("accept") String acceptHeader) {
        List<Client> clients = service.getAll();
        log.info(clients.size() + " items was retrieved from db, and response was send in " + acceptHeader + "format");
        return clients;
    }

}
